app/requirements.txt
attrs==22.1.0
biopython==1.79
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
exceptiongroup==1.0.4
fire==0.4.0
Flask==2.2.2
fonttools==4.38.0
greenlet==2.0.1
importlib-metadata==5.1.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.44
tomli==2.0.1
Werkzeug==2.2.2
zipp==3.10.0
